education-dashboard-api/requirements.txt
flask==2.2.3
flask-cors==3.0.10
werkzeug==2.2.3
pandas
python-dotenv>=0.21.0
openai>=1.68.2
numpy
tenacity>=8.1.0
PyPDF2==3.0.1
langchain>=0.3.21
pydantic>=2.5.2
sentence-transformers==2.2.2
huggingface-hub==0.16.4
langchain-community>=0.3.20
langchain-openai>=0.3.10
pdfplumber>=0.10.0
pycryptodome>=3.18.0
pymupdf
pikepdf